Fix depth size on ES3; Avoid ES2 using ES3 syms; -FBB on ES2; +ES support
SDL2 keeps returning addresses for GL functions regardless if they're properly
implemented or not. FNA now checks the ES context version and decides which
functions to use in affected cases.

ES2 lacks framebuffer blitting. For now, FNA forcibly disables it on ES2 contexts.
This just hides the crash and all warnings; an FBB for ES2 is still required.

Added support for:
* KHR_debug (alt. to ARB_debug_output)
* glColorMaskiOES from OES_draw_buffers_indexed (alt. to glColorMaskIndexedEXT from EXT_draw_buffers2)
* glDraw{RangeElements,ElementsInstanced}BaseVertexOES from OES_draw_elements_base_vertex if using ES3 and extension(!) available
* Other OES extensions where required